==> Codes/1-Getting All Websites.ipynb <==
# Import 
from splinter import Browser
from bs4 import BeautifulSoup
import pandas as pd
import requests
from webdriver_manager.chrome import ChromeDriverManager

# Set the executable path and initialize Splinter
executable_path = {'executable_path': ChromeDriverManager().install()}
browser = Browser('chrome', **executable_path, headless=False)

link_array = []
BASE_URL = "https://sh.lianjia.com/chengjiao/"
for i in range(1,101):     
    URL = BASE_URL + 'pg' + str(i)
    browser.visit(URL)
    r = requests.get(URL, timeout=None) 
    html = browser.html
    soup = BeautifulSoup(html, 'html.parser')
    links = soup.findAll('a', {'class':'img'})
    for k in links:
        link_array.append(k['href'])
browser.quit()

links = pd.DataFrame(link_array).drop_duplicates()
links.to_csv('../links for Deals in Shanghai.csv', index = False)
