==> Codes/2-Final Deals Information.ipynb <==
# Import 
from splinter import Browser
from bs4 import BeautifulSoup
import pandas as pd
import requests
from webdriver_manager.chrome import ChromeDriverManager

# Set the executable path and initialize Splinter
executable_path = {'executable_path': ChromeDriverManager().install()}
browser = Browser('chrome', **executable_path, headless=False)

def basic_attributes():
    source_info_basic=[]
    content_basic = soup.find('div', class_ = 'content')
    detail_basic = content_basic.find('ul')
    for li in detail_basic.find_all('li'):
        info_basic = li.text
        source_info_basic.append(info_basic)
    return source_info_basic


def deal_attributes():
    source_info_deal=[]
    content_deal = soup.find('div', class_ = 'transaction')
    detial_deal = content_deal.find('ul')
    for li in detial_deal.find_all('li'):
        info_deal = li.text
        source_info_deal.append(info_deal)
    return source_info_deal

def list_format():
    draft_information=[]
    source_info = basic_attributes() + deal_attributes()
    source_information = pd.DataFrame(source_info)
    for i in range(0, len(source_info)):
        draft_information.append(source_information[0][i][4:])
    return draft_information

def view_attributes():
    source_info_views=[]
    content_views = soup.find('div', class_ = 'msg')
    for span in content_views.find_all('span'):
        info_view = span.text
        source_info_views.append(info_view)
    return source_info_views

def list_format_with_view():
    final_information = list_format() + (view_attributes())
    return final_information

def deal_information():
    totalprice = soup.find('div', class_='price').span.text
    unitprice = soup.find('div', class_='price').b.text
    dealdate = soup.find('div', class_='wrapper').span.text
    title = soup.find('div', class_='wrapper').h1.text
    return [totalprice, unitprice, dealdate, title]

def get_final_info():
    a = list_format_with_view()
    b = deal_information()
    results = a + b
    return results

# Import Links to array 
links = pd.read_csv('Resources/Shanghai Deals Links.csv')
links_array = links.to_numpy()

# iterate
final= []
header = {'User-Agent':'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.130 Safari/537.36'}

for j in range(0,len(links_array)):
    URL = links_array[j][0]
    try: 
        browser.visit(URL) 
        r = requests.get(URL, headers = header, timeout=None)  
        html = browser.html
        soup = BeautifulSoup(html, 'html.parser')
        final.append(get_final_info())
        print(j)
    except Exception:
        print('Timeout')
browser.quit()

final = pd.DataFrame(final)
final.to_csv('../Resources/Shanghai Deals Information.csv', index = False)
