==> Codes/3-Final Selling Information.ipynb <==
# Import 
from splinter import Browser
from bs4 import BeautifulSoup
import pandas as pd
import requests
from webdriver_manager.chrome import ChromeDriverManager

# Set the executable path and initialize Splinter
executable_path = {'executable_path': ChromeDriverManager().install()}
browser = Browser('chrome', **executable_path, headless=False)

def basic():
    info = []
    attributes = soup.find('div', class_= 'introContent')
    basic = attributes.find('div', class_ = 'content')
    transaction = attributes.find('div', class_ = 'transaction') 
    ul_basic = basic.find('ul')
    ul_transaction = transaction.find('ul')
    for li_basic in ul_basic.find_all('li'):
        info_basic = li_basic.text[4:]
        info.append(info_basic)
    for li_transaction in ul_transaction.find_all('li'):
        info_transaction = li_transaction.text[5:]
        info.append(info_transaction)
        info = [x.strip() for x in info if x.strip() !='']
    return info

def price_and_year_built():
    totalprice = soup.find('div', class_='price').span.text
    unitprice = soup.find('div', class_ = 'unitPrice').span.text
    year_built = soup.find('div', class_ = 'subInfo noHidden').text
    return [totalprice, unitprice, year_built]

def name_and_area_code():
    around_information = soup.find('div', class_ = 'aroundInfo')
    name = around_information.find('div', class_ = 'communityName').a.text
    area = around_information.find('span', class_ = 'info').a.text
    code_information = soup.find('div', class_ = 'houseRecord')
    code = code_information.find('span', class_ ='info').text[0:12]
    return [name, area, code]


def merge():
    final_info = name_and_area_code() + price_and_year_built() + basic()
    return final_info

# Import Links to array 
links = pd.read_csv('Resources/Shanghai Selling Links.csv')
links_array = links.to_numpy()

# iterate
final= []
header = {'User-Agent':'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.130 Safari/537.36'}
for j in range(0,len(links_array)):
    URL = links_array[j][0]
    try:
        r = requests.get(URL, headers = header, timeout=None)  
        browser.visit(URL)
        html = browser.html
        soup = BeautifulSoup(html, 'html.parser')
        final.append(merge())
        print(j)
    except Exception:
        print('Timeout')
browser.quit()

final = pd.DataFrame(final)
final.to_csv('../Resources/Shanghai Selling Information.csv', index = False)
