==> Codes/7-Data Processing.ipynb <==
# Import
import pandas as pd
import sqlite3

# Read database files
con = sqlite3.connect('../HouseInfo.db')
deal = pd.read_sql_query('Select * from Deal', con)
selling = pd.read_sql_query('Select * from Selling', con)
pd.set_option('display.max_columns', None)

# Match Columns
# deal file
deal_order = ['City','HouseName','district','totalPrice','price','floor','square','constructionTime','bedRoom','livingRoom','kitchen','bathRoom','buildingType','buildingStructure','renovationCondition','elevator','subway','tradeTime','DOM','followers','Lng','Lat']
deal = deal[deal_order]
deal = deal.rename(columns={
    'HouseName': 'HouseName',
    'district': 'District',
    'totalPrice': 'TotalPrice',
    'price': 'Price',
    'floor': 'Floor',
    'square': 'Square',
    'constructionTime': 'ConstructionTime',
    'bedRoom': 'Bedrooms',
    'livingRoom': 'LivingRooms',
    'kitchen': 'Kitchens',
    'bathRoom': 'Bathrooms',
    'buildingType': 'Type',
    'buildingStructure': 'Structure',
    'renovationCondition': 'Renovation',
    'elevator': 'Elevator',
    'subway': 'Subway',
    'tradeTime': 'TradeTime',
    'DOM': 'DaysonMarket',
    'followers': 'Followers'
})

deal['Bedrooms'] = deal['Bedrooms'].replace(' ','0')
deal['LivingRooms'] = deal['Bedrooms'].replace(' ','0')
deal['Kitchens'] = deal['Kitchens'].replace(' ','0')
deal['Bathrooms'] = deal['Bathrooms'].replace(' ','0')
deal['Floor'] = deal['Floor'].replace('地下室','Low')

deal['City'] = deal['City'].astype(str)
deal['HouseName'] = deal['HouseName'].astype(str)
deal['District'] = deal['District'].astype(str)
deal['TotalPrice'] = deal['TotalPrice'].astype(str)
deal['Price'] = deal['Price'].astype(float)
deal['Floor'] = deal['Floor'].astype(str)
deal['Square'] = deal['Square'].astype(float)
deal['ConstructionTime'] = deal['ConstructionTime'].astype(str)
deal['Bedrooms'] = deal['Bedrooms'].astype(str)
deal['LivingRooms'] = deal['LivingRooms'].astype(str)
deal['Kitchens'] = deal['Kitchens'].astype(str)
deal['Bathrooms'] = deal['Bathrooms'].astype(str)
deal['Type'] = deal['Type'].astype(str)
deal['Structure'] = deal['Structure'].astype(str)
deal['Renovation'] = deal['Renovation'].astype(str)
deal['Elevator'] = deal['Elevator'].astype(str)
deal['TradeTime'] = deal['TradeTime'].astype(int)
deal['DaysonMarket'] = deal['DaysonMarket'].astype(int)
deal['Followers'] = deal['Followers'].astype(int)
deal['Lng'] = deal['Lng'].astype(float)
deal['Lat'] = deal['Lat'].astype(float)

# Match Columns
# selling file
selling_order = ['City','HouseName','District','TotalPrice','Price','Floor','Square','ConstructionTime','BedRooms','LivingRooms','Kitchen','BathRooms','BuildingType','BuildingStructure','RenovationCondition','Elevator','FloorType','Lng','Lat','id']
selling = selling[selling_order]
selling = selling.rename(columns={
    'BedRooms': 'Bedrooms',
    'Kitchen': 'Kitchens',
    'BathRooms': 'Bathrooms',
    'BuildingType': 'Type',
    'BuildingStructure': 'Structure',
    'RenovationCondition': 'Renovation',
    'FloorType': 'FloorType',
    'id' : 'ID'
})

selling['City'] = selling['City'].astype(str)
selling['HouseName'] = selling['HouseName'].astype(str)
selling['District'] = selling['District'].astype(str)
selling['Floor'] = selling['Floor'].astype(str)
selling['ConstructionTime'] = selling['ConstructionTime'].astype(str)
selling['Bedrooms'] = selling['Bedrooms'].astype(int)
selling['LivingRooms'] = selling['LivingRooms'].astype(int)
selling['Kitchens'] = selling['Kitchens'].astype(int)
selling['Bathrooms'] = selling['Bathrooms'].astype(int)
selling['Type'] = selling['Type'].astype(str)
selling['Structure'] = selling['Structure'].astype(str)
selling['Renovation'] = selling['Renovation'].astype(str)
selling['Elevator'] = selling['Elevator'].astype(str)
selling['FloorType'] = selling['FloorType'].astype(str)
selling['Lng'] = selling['Lng'].astype(float)
selling['Lat'] = selling['Lat'].astype(float)
selling['ID'] = selling['ID'].astype(str)

deal.to_csv('../Resources/Final Deal.csv', index=None)
selling.to_csv('../Resources/Final Selling.csv', index=None)
