==> Codes/9-Resize Deal for html.ipynb <==
# Import
import pandas as pd
import json
import csv

# Read File
deal = pd.read_csv('../Resources/Final Deal.csv', low_memory=False)
deal = pd.DataFrame(deal)
deal.head()

# Funtion 
def csv_to_json(csvFilePath, jsonFilePath):
    jsonArray = []
      
    #read csv file
    with open(csvFilePath, encoding='utf-8') as csvf: 
        #load csv file data using csv library's dictionary reader
        csvReader = csv.DictReader(csvf) 

        #convert each csv row into python dict
        for row in csvReader: 
            #add this python dict to json array
            jsonArray.append(row)
  
    #convert python jsonArray to JSON String and write to file
    with open(jsonFilePath, 'w', encoding='utf-8') as jsonf: 
        jsonString = json.dumps(jsonArray,ensure_ascii=False, indent=4)
        jsonf.write(jsonString)

len(deal)

deal = deal.drop(labels=range(5000,159000),axis=0)

len(deal)

deal.head()

deal.to_csv('../Resources/Deal_html.csv', index=False)

# Deal.json
csvFilePath = '../Resources/Deal_html.csv'
jsonFilePath = '../Resources/Final Deal_html.json'
csv_to_json(csvFilePath, jsonFilePath)
